pyproject.toml
[tool.poetry]
name = "thecatapi"
version = "0.1.0"
description = ""
authors = ["Quesia Santos <quesiasts@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
requests = "^2.28.1"
Django = "^4.1"
zappa = "^0.55.0"
awscli = "*"
uritemplate = "^4.1.1"
pyyaml = ">=5.4,<6.0"
djangorestframework = "^3.13.1"
prettyconf = "^2.2.1"
psycopg2 = "^2.9.3"
pytest = "^7.1.2"
sentry-sdk = "^1.9.5"
Pillow = "^9.2.0"
pre-commit = "^2.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
